Breaking change: VNode children method changed
- VNode children method is now accepting only arrays.
- Added new VNode method "child" that accepts string or single VNode.

Reduce number of allocations when there is a single VNode child, I've
found that is is quite important in some cases with large lists.

It is implemented this way to reduce number of state transitions like
"string" => "array", "array" => "string". It makes API uglier, but at
the same time it won't allow developers to write even more uglier code
that abuses such transitions.